Improved 'audio' code.
Fixed mess with types conversion.
Added comments.

+++ src/core/audio/music.cpp
@@ -152,7 +152,9 @@ bool sMusic::Update(uint32_t CurrentTick)
 
 	if (FadeInSwitch) {
 		FadeTicks += TicksDelta;
-		LocalVolume = FadeEndVol * FadeTicks / FadePeriod;
+		// we are safe with static_cast here, since FadeTicks and FadePeriod
+		// will not exceed 'float' in our case for sure (usualy <10000 ticks)
+		LocalVolume = FadeEndVol * static_cast<float>(FadeTicks) / static_cast<float>(FadePeriod);
 		if (LocalVolume >= FadeEndVol) {
 			LocalVolume = FadeEndVol;
 			FadeInSwitch = false;
@@ -163,7 +165,9 @@ bool sMusic::Update(uint32_t CurrentTick)
 
 	if (FadeOutSwitch) {
 		FadeTicks += TicksDelta;
-		LocalVolume = 1.0f - FadeStartVol * FadeTicks / FadePeriod;
+		// we are safe with static_cast here, since FadeTicks and FadePeriod
+		// will not exceed 'float' in our case for sure (usualy <10000 ticks)
+		LocalVolume = 1.0f - FadeStartVol * static_cast<float>(FadeTicks) / static_cast<float>(FadePeriod);
 		if (LocalVolume < 0.0f) {
 			LocalVolume = 0.0f;
 			FadeOutSwitch = false;
