+++ src/core/audio/sound.cpp
@@ -300,9 +300,12 @@ void vw_UpdateSound(uint32_t CurrentTick)
 			iter->second.DestroyTicks += DeltaTicks;
 
 			if (iter->second.DestroyTicks < iter->second.DestroyPeriod) {
+				// we are safe with static_cast here, since DestroyTicks and DestroyPeriod
+				// will not exceed 'float' in our case for sure (usualy <10000 ticks)
 				alSourcef(iter->second.Source, AL_GAIN,
 					  iter->second.GlobalVolume * iter->second.LocalVolume *
-					  (1.0f - iter->second.DestroyTicks / iter->second.DestroyPeriod));
+					  (1.0f - static_cast<float>(iter->second.DestroyTicks) /
+						  static_cast<float>(iter->second.DestroyPeriod)));
 				alGetError(); // reset errors
 			} else {
 				// release, volume less or equal zero
